feat: optimize utils and extend types for storage
