+++ crud/common/CrudAddEdit.tsx
@@ -21,7 +21,7 @@ import { TBaseParameter, TResponse, TUnit } from "@arkejs/client";
 import { Button, Dialog, Spinner } from "@arkejs/ui";
 import toast from "react-hot-toast";
 import { useRouter } from "next/router";
-import { containsFile } from "../../utils/file";
+import { containsFile } from "@/utils/file";
 
 export interface CrudProps {
   unitId?: string;
@@ -52,12 +52,9 @@ export function CrudAddEdit(props: CrudProps) {
         });
       }
       const payload = containsFile(data) ? formData : data;
-      // TODO: require types extension on client
       const promise = !unitId
-        ? // @ts-ignore
-          client.unit.create(arkeId, payload)
-        : // @ts-ignore
-          client.unit.edit(arkeId, unitId as string, payload);
+        ? client.unit.create(arkeId, payload)
+        : client.unit.edit(arkeId, unitId as string, payload);
       promise.then(
         (res) => onSubmit(res),
         (err) =>
