+++ tsconfig.json
@@ -23,7 +23,8 @@
       "@/pages/*": ["pages/*"],
       "@/server/*": ["server/*"],
       "@/styles/*": ["styles/*"],
-      "@/types/*": ["types/*"]
+      "@/types/*": ["types/*"],
+      "@/utils/*": ["utils/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
